examples/pytorch/image-text/requirements.txt
accelerate==0.24.1
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.1.0
bitsandbytes==0.41.2.post2
certifi==2023.11.17
charset-normalizer==3.3.2
datasets==2.15.0
dill==0.3.7
evaluate==0.4.1
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
huggingface-hub==0.19.4
idna==3.6
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
packaging==23.2
pandas==2.1.3
peft==0.6.2
Pillow==10.1.0
protobuf==4.25.1
psutil==5.9.6
pyarrow==14.0.1
pyarrow-hotfix==0.6
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
responses==0.18.0
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
sentencepiece==0.1.99
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
transformers @ git+https://github.com/huggingface/transformers.git@0864dd3beb238b7bec3528a3d1d6c17a28f51a51
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
xxhash==3.4.1
yarl==1.9.3
